+++ packages/spacecat-shared-data-access/src/util/index.js
@@ -28,6 +28,14 @@ export {
   resetLoggerRegistry,
 } from './logger-registry.js';
 
+export {
+  URL_INDEX_TABLES,
+  URL_CHUNK_SIZE,
+  syncUrlIndex,
+  syncUrlIndexMany,
+  lookupEntityIdsByUrl,
+} from './url-index.utils.js';
+
 /**
  * Datastore types that collections can use to declare their storage backend.
  * @readonly